day 1 2023 solution
Did not go for an efficient solution since the problem seems very
simple.

Did have to debug the fact that oneight should be 1 for min and
8 for max case, so simple string substitution fails. I could
have still done string substitution with indexes but at that point
it was easier to use find and rfind to find the first and last
index of any number, and use that number as the first or last
if the index isless than the char based one.