day2
This day made me miss easy batch iteration (can do it
with zip etc., but is not convenient).

An easier way was to use pandas/numpy and just reshape
the array into a 2D one in numpy or do it appropriately
in pandas.

A final way was just to iterate and grab each number first
then the cube name.

I chose to use chunked, because it has been implemented
as itertools.batched in the standard library in python 3.12.

Change-Id: I32eac3ef1702b3c515129a0ca33757b5491cc133