day 3
This can be considered a variant of grid graphs, however
the first part was solveable without having to set up a grid
graph, using breadth first search, leveraging the fact that
numbers only read left to right and are on the same line.
Just grab the connected status of the part by scanning all connected
directions for a symbol while you parse the part number. Only tricky
thing was to ensure recording the part number when we moved to a new
row as well as during the iteration of the columns.

For part 2, I used a grid graph solution once again using breadth
first search, but only searching the connected points whenever we
reached a * while parsing the schematic.

A more general solution setting up a grid graph of neighbor mappings
for all numbers and symbols was possible, but not needed. Still,
this took a while to implement and recall BFS and grid graph basics,
so the general solution may have been simpler from the beginning..

Change-Id: I708cc5fd828ce673ee0df2ff490307368b72d3e6