day 4 2023, stack
Day 4 was simple, create a stack, and process winning
cards on the stack. Part 1 is just 2^(n-1) points
for n winning numbers per card, iterating once.

part 2 is best solved by adding winning cards to the stack.
I used a deque so popleft was just as efficient as pop right
and appended winning cards as we came upon each or each copy.

Change-Id: I64eabf8e814e8dbe454e1f325c0609fd2aa8f71b