Solve day 10
For the first part I created a graph and just traversed from the
starting point.

For the second part, I created a `shapely` polygon as I traversed
the graph from the points. Then I "prepared" the polygon to make
it much more efficient for repeat calculation. This takes the
run time down from several minutes to fractions of a second.
Finally I used numpy vectorized approaches, and brute force
found each point that lies within the polygon.

Other approaches I thought of are ray casting, monte-carlo sampling
for area (using shapely again), or the shoe-string/pencil method,
whatever it is called. I will look it up after. Basically you check
how many times you cross the line of the polygon. That is probably
the intended way of solving it, but this was fast and quick.

Change-Id: I7dd9c0285457a255e3b7ff9ad7b6004ffba2f40c